split TextFieldLabel.spec into LabelBase.spec
